tools/requirements.txt
Pillow>=10.0.0
piexif>=1.1.3
PyYAML>=6.0.1
requests>=2.25.0
